Refactor configuration loading and database setup
Switched from using a TOML file to a properties file for configuration loading. Updated the database setup process to use these new properties to set up a HikariDataSource, then used this data source to start a StaticSaduLoader. Also moved the initialization of the MessageBuilder from "config.toml" to "config.properties".