Add Message class and improve Language and MessageBuilderImpl classes
Added a new Message class and implemented a new method in the Language class to get the messages from the database asynchronously. The MessageBuilderImpl class was updated to store these locale-specific messages in a Map object for quick and easy access. This refactoring aims to improve performance and simplify the handling of localized messages.