Message fetching (#20)
* Add Message class and improve Language and MessageBuilderImpl classes

Added a new Message class and implemented a new method in the Language class to get the messages from the database asynchronously. The MessageBuilderImpl class was updated to store these locale-specific messages in a Map object for quick and easy access. This refactoring aims to improve performance and simplify the handling of localized messages.

* Refactor MessageBuilderTest and enhance localization functionality (#19)

Refactored MessageBuilderTest, removing redundant server object and using explicit method calls on MessageBuilder. Improved localization by storing localized messages in a Map within the MessageBuilderImpl class. Added two new tests for validating the absence of German messages and presence of English messages. The aim is to enhance performance and provide simpler management of localized messages.